fix: harvest all button should disabled when no earning
